Reworked the wrapper interface
