+++ Basic.Azure.Storage.Tests.Integration/BlobServiceClientTests.cs
@@ -22,52 +22,8 @@ using LeaseStatus = Basic.Azure.Storage.Communications.Common.LeaseStatus;
 namespace Basic.Azure.Storage.Tests.Integration
 {
     [TestFixture]
-    public class BlobServiceClientTests
+    public class BlobServiceClientTests : BaseBlobServiceClientTestFixture
     {
-
-        private readonly StorageAccountSettings _accountSettings = new LocalEmulatorAccountSettings();
-        private readonly CloudStorageAccount _storageAccount = CloudStorageAccount.Parse("UseDevelopmentStorage=true");
-
-        private readonly Dictionary<string, string> _containersToCleanUp = new Dictionary<string, string>();
-
-        private string GenerateSampleContainerName()
-        {
-            var name = "unit-test-" + Guid.NewGuid().ToString().ToLower();
-            RegisterContainerForCleanup(name, null);
-            return name;
-        }
-
-        private string GenerateSampleBlobName()
-        {
-            return String.Format("unit-test-{0}", Guid.NewGuid());
-        }
-
-        private void RegisterContainerForCleanup(string containerName, string leaseId)
-        {
-            _containersToCleanUp[containerName] = leaseId;
-        }
-
-        public string FakeLeaseId { get { return "a28cf439-8776-4653-9ce8-4e3df49b4a72"; } }
-
-        [TestFixtureTearDown]
-        public void TestFixtureTeardown()
-        {
-            //let's clean up!
-            var client = _storageAccount.CreateCloudBlobClient();
-            foreach (var containerPair in _containersToCleanUp)
-            {
-                var container = client.GetContainerReference(containerPair.Key);
-                if (!string.IsNullOrEmpty(containerPair.Value))
-                {
-                    try
-                    {
-                        container.ReleaseLease(new AccessCondition() { LeaseId = containerPair.Value });
-                    }
-                    catch { }
-                }
-                container.DeleteIfExists();
-            }
-        }
 
         #region Account Operations
 
@@ -2543,366 +2499,5 @@ namespace Basic.Azure.Storage.Tests.Integration
 
         #endregion
 
-        #region Assertions
-
-        private void AssertContainerExists(string containerName)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            if (!container.Exists())
-                Assert.Fail(String.Format("AssertContainerExists: The container '{0}' does not exist", containerName));
-        }
-
-        private void AssertContainerDoesNotExist(string containerName)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            if (container.Exists())
-                Assert.Fail(String.Format("AssertContainerDoesNotExist: The container '{0}' exists", containerName));
-        }
-
-        private void AssertContainerIsLeased(string containerName, string leaseId)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            if (!container.Exists())
-                Assert.Fail(String.Format("AssertContainerIsLeased: The container '{0}' does not exist", containerName));
-
-            try
-            {
-                container.RenewLease(new AccessCondition() { LeaseId = leaseId });
-            }
-            catch (Exception exc)
-            {
-                Assert.Fail(String.Format("AssertContainerIsLeased: The container '{0}' gave an {1} exception when renewing with the specified lease id: {2}", containerName, exc.GetType().Name, exc.Message));
-            }
-        }
-
-        private void AssertContainerIsNotLeased(string containerName)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            if (!container.Exists())
-                Assert.Fail(String.Format("AssertContainerIsNotLeased: The container '{0}' does not exist", containerName));
-
-            try
-            {
-                var leaseId = container.AcquireLease(null, null);
-                RegisterContainerForCleanup(containerName, leaseId);
-            }
-            catch (Exception exc)
-            {
-                Assert.Fail(String.Format("AssertContainerIsNotLeased: The container '{0}' gave an {1} exception when attempting to acquire a new lease: {2}", containerName, exc.GetType().Name, exc.Message));
-            }
-        }
-
-        private void AssertContainerAccess(string containerName, Microsoft.WindowsAzure.Storage.Blob.BlobContainerPublicAccessType containerAccessType)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            if (!container.Exists())
-                Assert.Fail(String.Format("AssertContainerAccess: The container '{0}' does not exist", containerName));
-
-            var permissions = container.GetPermissions();
-            Assert.AreEqual(containerAccessType, permissions.PublicAccess, String.Format("AssertContainerAccess: Container access was expected to be {0}, but it is actually {1}", containerAccessType, permissions.PublicAccess));
-        }
-
-        private Microsoft.WindowsAzure.Storage.Blob.ICloudBlob AssertBlobExists(string containerName, string blobName, BlobType blobType)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            if (!container.Exists())
-                Assert.Fail("AssertBlobExists: The container '{0}' does not exist", containerName);
-
-            var blob = (blobType == BlobType.BlockBlob
-                ? (ICloudBlob)container.GetBlockBlobReference(blobName)
-                : (ICloudBlob)container.GetPageBlobReference(blobName));
-
-            if (!blob.Exists())
-                Assert.Fail("AssertBlobExists: The blob '{0}' does not exist", blobName);
-
-            return blob;
-        }
-
-        private Microsoft.WindowsAzure.Storage.Blob.ListBlockItem AssertBlockExists(string containerName, string blobName, string blockId, BlockListingFilter blockType = BlockListingFilter.All)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            if (!container.Exists())
-                Assert.Fail("AssertBlockExists: The container '{0}' does not exist", containerName);
-
-            var blob = container.GetBlockBlobReference(blobName);
-            var blockList = blob.DownloadBlockList(blockType);
-            var block = blockList.FirstOrDefault(item => item.Name == blockId);
-
-            if (block == null)
-                Assert.Fail("AssertBlockExists: The block of id '{0}' does not exist", blockId);
-
-            return block;
-        }
-
-        private Microsoft.WindowsAzure.Storage.Blob.ICloudBlob AssertBlobDoesNotExist(string containerName, string blobName, BlobType blobType)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            if (!container.Exists())
-                Assert.Fail("AssertBlobDoesNotExist: The container '{0}' does not exist", containerName);
-
-            var blob = (blobType == BlobType.BlockBlob
-                ? (ICloudBlob)container.GetBlockBlobReference(blobName)
-                : (ICloudBlob)container.GetPageBlobReference(blobName));
-
-            if (blob.Exists())
-                Assert.Fail("AssertBlobDoesNotExist: The blob '{0}' exists", blobName);
-
-            return blob;
-
-        }
-
-        private Microsoft.WindowsAzure.Storage.Blob.ICloudBlob AssertBlobContainsData(string containerName, string blobName, BlobType blobType, byte[] expectedData)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            if (!container.Exists())
-                Assert.Fail("AssertBlobContainsData: The container '{0}' does not exist", containerName);
-
-            var blob = (blobType == BlobType.BlockBlob
-                ? (ICloudBlob)container.GetBlockBlobReference(blobName)
-                : (ICloudBlob)container.GetPageBlobReference(blobName));
-
-            if (!blob.Exists())
-                Assert.Fail("AssertBlobContainsData: The blob '{0}' does not exist", blobName);
-
-            using (var stream = new MemoryStream())
-            {
-                blob.DownloadToStream(stream);
-
-                var gottenData = stream.ToArray();
-
-                // Comparing strings -> MUCH faster than comparing the raw arrays
-                var gottenDataString = Convert.ToBase64String(gottenData);
-                var expectedDataString = Convert.ToBase64String(expectedData);
-
-                Assert.AreEqual(expectedData.Length, gottenData.Length);
-                Assert.AreEqual(gottenDataString, expectedDataString);
-            }
-
-            return blob;
-        }
-
-        private IDictionary<string, string> GetContainerMetadata(string containerName)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-
-            container.FetchAttributes();
-            return container.Metadata;
-        }
-
-        private void AssertIdentifierInSharedAccessPolicies(Microsoft.WindowsAzure.Storage.Blob.SharedAccessBlobPolicies sharedAccessPolicies, BlobSignedIdentifier expectedIdentifier, Microsoft.WindowsAzure.Storage.Blob.SharedAccessBlobPermissions permissions)
-        {
-            var policy = sharedAccessPolicies.Where(i => i.Key.Equals(expectedIdentifier.Id, StringComparison.InvariantCultureIgnoreCase)).FirstOrDefault();
-            Assert.IsNotNull(policy);
-            Assert.AreEqual(expectedIdentifier.AccessPolicy.StartTime, policy.Value.SharedAccessStartTime.Value.UtcDateTime);
-            Assert.AreEqual(expectedIdentifier.AccessPolicy.Expiry, policy.Value.SharedAccessExpiryTime.Value.UtcDateTime);
-            Assert.IsTrue(policy.Value.Permissions.HasFlag(permissions));
-        }
-
-        #endregion
-
-        #region Setup Mechanics
-
-        private void CreateContainer(string containerName, Dictionary<string, string> metadata = null)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-
-            container.Create();
-
-            if (metadata != null)
-            {
-                // why ???
-                foreach (var key in metadata.Keys)
-                {
-                    container.Metadata.Add(key, metadata[key]);
-                }
-                container.SetMetadata();
-            }
-
-        }
-
-        private string LeaseContainer(string containerName, TimeSpan? leaseTime, string leaseId)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            var lease = container.AcquireLease(leaseTime, leaseId);
-            RegisterContainerForCleanup(containerName, lease);
-            return lease;
-        }
-
-        private void ReleaseContainerLease(string containerName, string lease)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            container.ReleaseLease(new AccessCondition() { LeaseId = lease });
-        }
-
-        private void BreakContainerLease(string containerName, string lease)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            container.BreakLease(TimeSpan.FromSeconds(1), new AccessCondition() { LeaseId = lease });
-        }
-
-        private void AddContainerAccessPolicy(string containerName, Microsoft.WindowsAzure.Storage.Blob.BlobContainerPublicAccessType publicAccess, string id = null, DateTime? startDate = null, DateTime? expiry = null)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            var permissions = container.GetPermissions();
-            permissions.PublicAccess = publicAccess;
-            if (!string.IsNullOrEmpty(id) && startDate.HasValue && expiry.HasValue)
-            {
-                permissions.SharedAccessPolicies.Add(id, new Microsoft.WindowsAzure.Storage.Blob.SharedAccessBlobPolicy()
-                {
-                    Permissions = Microsoft.WindowsAzure.Storage.Blob.SharedAccessBlobPermissions.Read,
-                    SharedAccessStartTime = startDate,
-                    SharedAccessExpiryTime = expiry
-                });
-            }
-            container.SetPermissions(permissions);
-        }
-
-        private Microsoft.WindowsAzure.Storage.Blob.BlobContainerPermissions GetContainerPermissions(string containerName)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            var permissions = container.GetPermissions();
-            return permissions;
-        }
-
-        private Microsoft.WindowsAzure.Storage.Blob.LeaseState GetContainerLeaseState(string containerName)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            container.FetchAttributes();
-            return container.Properties.LeaseState;
-        }
-
-        private List<Microsoft.WindowsAzure.Storage.Blob.ListBlockItem> CreateBlockList(string containerName, string blobName,
-            IEnumerable<string> blockIdsToCreate, string dataPerBlock, Encoding encoder = null)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            var blob = container.GetBlockBlobReference(blobName);
-
-            byte[] data = (encoder ?? Encoding.Unicode).GetBytes(dataPerBlock);
-            // non-Base64 values fail?
-            foreach (var curBlockId in blockIdsToCreate)
-            {
-                blob.PutBlock(curBlockId, new MemoryStream(data), null);
-            }
-
-            return blob
-                .DownloadBlockList(BlockListingFilter.All)
-                .ToList();
-        }
-
-        private BlockListBlockIdList CreateBlockIdList(int idCount, BlockListListType listType)
-        {
-            var idList = new BlockListBlockIdList();
-            for (var i = 0; i < idCount; i++)
-            {
-                idList.Add(new BlockListBlockId
-                {
-                    Id = Base64Converter.ConvertToBase64("id" + idCount),
-                    ListType = listType
-                });
-            }
-            return idList;
-        }
-
-        private List<string> GetIdsFromBlockIdList(BlockListBlockIdList list)
-        {
-            return list.Select(bid => bid.Id).ToList();
-        }
-
-        private void CreateBlockBlob(string containerName, string blobName, Dictionary<string, string> metadata = null, string content = "Generic content")
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            var blob = container.GetBlockBlobReference(blobName);
-
-            byte[] data = UTF8Encoding.UTF8.GetBytes(content);
-            blob.UploadFromByteArray(data, 0, data.Length);
-
-            if (metadata != null)
-            {
-                foreach (var key in metadata.Keys)
-                {
-                    blob.Metadata.Add(key, metadata[key]);
-                }
-                blob.SetMetadata();
-            }
-        }
-
-        private void CreatePageBlob(string containerName, string blobName, Dictionary<string, string> metadata = null)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            var blob = container.GetPageBlobReference(blobName);
-
-            byte[] data = UTF8Encoding.UTF8.GetBytes("Generic content");
-            var necessarySize = data.Length + (512 - (data.Length % 512));
-            var pageData = new byte[necessarySize];
-            data.CopyTo(pageData, 0);
-            blob.UploadFromByteArray(pageData, 0, pageData.Length);
-
-            if (metadata != null)
-            {
-                foreach (var key in metadata.Keys)
-                {
-                    blob.Metadata.Add(key, metadata[key]);
-                }
-                blob.SetMetadata();
-            }
-        }
-
-        private void CreateBlobUncommitted(string containerName, string blobName)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            var blob = container.GetBlockBlobReference(blobName);
-
-            byte[] data = UTF8Encoding.UTF8.GetBytes("Generic content");
-            // non-Base64 values fail?
-            var blockId = Convert.ToBase64String(Encoding.UTF8.GetBytes("A"));
-            blob.PutBlock(blockId, new MemoryStream(data), null);
-        }
-
-        private void CopyBlob(string containerName, string sourceBlobName, string targetBlobName)
-        {
-            // could we have made this require more work?
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            var sourceBlob = container.GetBlockBlobReference(sourceBlobName);
-            var targetBlob = container.GetBlockBlobReference(targetBlobName);
-            targetBlob.StartCopyFromBlob(sourceBlob);
-        }
-
-        private void SnapshotBlob(string containerName, string blobName)
-        {
-            var client = _storageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(containerName);
-            var blob = container.GetBlockBlobReference(blobName);
-            blob.CreateSnapshot();
-        }
-
-
-        #endregion
-
-        private DateTime GetTruncatedUtcNow()
-        {
-            return new DateTime(DateTime.UtcNow.Year, DateTime.UtcNow.Month, DateTime.UtcNow.Day, DateTime.UtcNow.Hour, DateTime.UtcNow.Minute, DateTime.UtcNow.Second, DateTimeKind.Utc);
-        }
     }
 }
